Remove IP address strorage from client player pool
+ Fix return value conversion warning at Util_GetTime when compiling.

+++ client/runutil.cpp
@@ -296,7 +296,7 @@ unsigned long Util_GetTime()
 		bInited = true;
 	}
 	QueryPerformanceCounter(&cli);
-	return (cli.QuadPart * 1000 / fli.QuadPart);
+	return (unsigned long)(cli.QuadPart * 1000 / fli.QuadPart);
 }
 
 //----------------------------------------------------
